Moved folder and removed eslint comments
